testsite/requirements.txt
django
django-simple-captcha
django-debug-toolbar
django-ckeditor
django-mptt